Add additional perm check for deletion
It is not ideal for the template (front-end) to use a different
permission check than the backend. In order to guarantee parity, the
check used for the client is also used on the server. Since these
functions use object_caches under the hood, this will not be costly for performance.

+++ opentreemap/treemap/views/map_feature.py
@@ -29,6 +29,7 @@ from treemap.lib.map_feature import (get_map_feature_or_404,
                                      context_dict_for_plot,
                                      context_dict_for_resource,
                                      context_dict_for_map_feature)
+from treemap.lib.perms import map_feature_is_deletable
 from treemap.views.misc import add_map_info_to_context
 
 
@@ -136,6 +137,9 @@ def update_map_feature_detail(request, instance, feature_id):
 
 def delete_map_feature(request, instance, feature_id):
     feature = get_map_feature_or_404(feature_id, instance)
+    if not map_feature_is_deletable(request.user.get_instance_user(instance),
+                                    feature):
+        raise ValidationError("map feature not deletable")
     feature.delete_with_user(request.user)
     return {'ok': True}
